requirements.txt
# Core
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=0.21.0
pydantic>=1.10.0
loguru>=0.6.0

# Traitement asynchrone
aiohttp>=3.8.0
asyncio>=3.4.3
aiofiles>=23.1.0

# Traitement PDF/Text
PyPDF2>=3.0.0
python-multipart>=0.0.5
python-magic>=0.4.27

# Utilitaires
tqdm>=4.65.0
requests>=2.28.0
python-json-logger>=2.0.7

# Cache et performances
aiocache>=0.12.0
orjson>=3.8.0

# Monitoring et métriques
prometheus-client>=0.16.0
opentelemetry-api>=1.16.0
opentelemetry-sdk>=1.16.0

# Tests et qualité
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
mypy>=1.2.0
flake8>=6.0.0
black>=23.3.0

# Développement
types-requests>=2.28.11
types-python-dateutil>=2.8.19
types-pyyaml>=6.0.12
